Fix confusing persistence == model && 'domain-specific' jargon
1.  Students were asking what a "domain-specific
model" is.  Tha magic adjective is rarely used in
teaching and its presence here is confusing e.g.
"What's a domain-specific model versus a
non-domain-specific model?"
2.  I don't consider the persistence interface
_really_ to be a Model.  In the sense that it's
neither a Controller nor a View, then yes.  And
that it's not an ORM-backed model versus a model,
OK, but it's really confusing at this stage of
their learning